wgpu: Inlined CommandRenderer::execute into Surface::draw_commands
